refactor(web): split Library into pages/library/ + shared components, fully tokenized
Break the 897-line Library.tsx into focused modules (index, sidebar, collection-bar,
share-collection-dialog, artifact-card, types) and convert every sub-component off
legacy CSS onto Tailwind tokens + ui/ primitives:
- Sidebar: rail + drawer + responsive states as tokenized utilities, accessible names
- CollectionBar: Button/Input variants
- ShareCollectionDialog: rebuilt on the Radix Dialog primitive (focus trap, esc, overlay)
- Extract Thumb, EmptyState, Spinner, RoleSelect into components/shared/ for reuse
All four themes + mobile drawer verified; typecheck + biome clean.
