+++ apps/web/src/routes/index.tsx
@@ -1,5 +1,5 @@
 import { createFileRoute } from "@tanstack/react-router"
-import { Library } from "../pages/Library"
+import { Library } from "../pages/library"
 
 export const Route = createFileRoute("/")({
   component: Library,
